patched changes from https://github.com/mobxjs/mobx-react-lite/pull/332/files
