+++ .gitignore
@@ -1,4 +1,5 @@
 /node_modules/
+/.idea/
 
 lib/
 es/
