+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nikhilthota1/mobx-react-lite",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "Lightweight React bindings for MobX based on React 16.8+ and Hooks",
   "main": "lib/index.js",
   "unpkg": "dist/mobxreactlite.umd.production.min.js",
@@ -61,7 +61,7 @@
     "@testing-library/react": "9.3.0",
     "@testing-library/react-hooks": "1.1.0",
     "@types/jest": "24.0.19",
-    "@types/node": "12.7.12",
+    "@types/node": "14.14.6",
     "@types/react": "16.9.6",
     "@types/react-dom": "16.9.2",
     "@typescript-eslint/eslint-plugin": "^2.19.2",
