+++ test/utils.ts
@@ -10,3 +10,9 @@ export function enableDevEnvironment() {
         process.env.NODE_ENV === "production"
     }
 }
+
+export function sleep(time: number) {
+    return new Promise<void>(res => {
+        setTimeout(res, time)
+    })
+}
